+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 - Use custom fork of Il2CppInterop
-- 
+- Added new way to hook methods (via `Hooks` nested type)
+- Added `cancelblueprints [radius]` command to cancel all blueprints in a radius
+- Support for mouse and controller input bindings